Proper data size for file decoding
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/pocketsphinx@12660 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ include/pocketsphinx.h
@@ -296,7 +296,7 @@ char *ps_lookup_word(ps_decoder_t *ps,
  * @return Number of samples of audio.
  */
 POCKETSPHINX_EXPORT
-int ps_decode_raw(ps_decoder_t *ps, FILE *rawfh,
+long ps_decode_raw(ps_decoder_t *ps, FILE *rawfh,
                   char const *uttid, long maxsamps);
 
 /**
